Moved files under com.google.cloud.tools.eclipse.appengine.facets.standard and com.google.cloud.tools.eclipse.appengine.facets.flex to com.google.cloud.tools.eclipse.appengine.facets
