+++ plugins/com.google.cloud.tools.eclipse.appengine.facets.test/src/com/google/cloud/tools/eclipse/appengine/facets/flex/AppEngineFlexFacetTest.java
@@ -4,6 +4,8 @@ import org.eclipse.wst.common.project.facet.core.ProjectFacetsManager;
 import org.junit.Assert;
 import org.junit.Test;
 
+import com.google.cloud.tools.eclipse.appengine.facets.AppEngineFlexFacet;
+
 public class AppEngineFlexFacetTest {
   @Test
   public void testFlexFacetExists() {
