+++ plugins/com.google.cloud.tools.eclipse.appengine.facets.test/src/com/google/cloud/tools/eclipse/appengine/facets/standard/AppEngineStandardFacetTest.java
@@ -24,7 +24,7 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
-import com.google.cloud.tools.eclipse.appengine.facets.standard.AppEngineStandardFacet;
+import com.google.cloud.tools.eclipse.appengine.facets.AppEngineStandardFacet;
 
 @RunWith(MockitoJUnitRunner.class)
 public class AppEngineStandardFacetTest {
