+++ plugins/com.google.cloud.tools.eclipse.appengine.facets/plugin.xml
@@ -22,13 +22,13 @@
       id="com.google.cloud.tools.eclipse.appengine.facet.install.action"
       facet="com.google.cloud.tools.eclipse.appengine.facet"
       type="INSTALL">
-      <delegate class="com.google.cloud.tools.eclipse.appengine.facets.standard.StandardFacetInstallDelegate"/>
+      <delegate class="com.google.cloud.tools.eclipse.appengine.facets.StandardFacetInstallDelegate"/>
     </action>
     <action
       id="com.google.cloud.tools.eclipse.appengine.facet.uninstall.action"
       facet="com.google.cloud.tools.eclipse.appengine.facet"
       type="UNINSTALL">
-      <delegate class="com.google.cloud.tools.eclipse.appengine.facets.standard.StandardFacetUninstallDelegate"/>
+      <delegate class="com.google.cloud.tools.eclipse.appengine.facets.StandardFacetUninstallDelegate"/>
     </action>
   </extension>
   
@@ -41,7 +41,7 @@
 
   <extension point="org.eclipse.wst.common.project.facet.core.listeners">
     <listener
-      class="com.google.cloud.tools.eclipse.appengine.facets.standard.AppEngineRuntimeChangeListener"
+      class="com.google.cloud.tools.eclipse.appengine.facets.AppEngineRuntimeChangeListener"
       eventTypes="PRIMARY_RUNTIME_CHANGED">
     </listener>
   </extension>
@@ -67,13 +67,13 @@
       id="com.google.cloud.tools.eclipse.appengine.facets.flex.install.action"
       facet="com.google.cloud.tools.eclipse.appengine.facets.flex"
       type="INSTALL">
-      <delegate class="com.google.cloud.tools.eclipse.appengine.facets.flex.FlexFacetInstallDelegate"/>
+      <delegate class="com.google.cloud.tools.eclipse.appengine.facets.FlexFacetInstallDelegate"/>
     </action>
     <action
       id="com.google.cloud.tools.eclipse.appengine.facets.flex.uninstall.action"
       facet="com.google.cloud.tools.eclipse.appengine.facets.flex"
       type="UNINSTALL">
-      <delegate class="com.google.cloud.tools.eclipse.appengine.facets.flex.FlexFacetUninstallDelegate"/>
+      <delegate class="com.google.cloud.tools.eclipse.appengine.facets.FlexFacetUninstallDelegate"/>
     </action>
   </extension>
