+++ plugins/com.google.cloud.tools.eclipse.appengine.facets/src/com/google/cloud/tools/eclipse/appengine/facets/AppEngineFlexFacet.java
@@ -1,4 +1,4 @@
-package com.google.cloud.tools.eclipse.appengine.facets.flex;
+package com.google.cloud.tools.eclipse.appengine.facets;
 
 public class AppEngineFlexFacet {
   public static final String ID = "com.google.cloud.tools.eclipse.appengine.facets.flex";
