+++ plugins/com.google.cloud.tools.eclipse.appengine.facets/src/com/google/cloud/tools/eclipse/appengine/facets/FlexFacetUninstallDelegate.java
@@ -1,4 +1,4 @@
-package com.google.cloud.tools.eclipse.appengine.facets.flex;
+package com.google.cloud.tools.eclipse.appengine.facets;
 
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.runtime.CoreException;
@@ -6,7 +6,7 @@ import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.wst.common.project.facet.core.IDelegate;
 import org.eclipse.wst.common.project.facet.core.IProjectFacetVersion;
 
-public class FlexFacetInstallDelegate implements IDelegate {
+public class FlexFacetUninstallDelegate implements IDelegate {
 
   @Override
   public void execute(IProject project, IProjectFacetVersion fv, Object config, IProgressMonitor monitor)
